updata

+++ src/main/java/com/coinwind/bifeng/ui/login/bean/LoginBean.java
@@ -175,7 +175,7 @@ public class LoginBean {
             private Object gz_type;
             private Object logo_url;
             private Object openid;
-            private int current_css;
+            private double current_css;
             private String head_img;
             private Object gz_name;
             private String type;
@@ -233,11 +233,11 @@ public class LoginBean {
                 this.openid = openid;
             }
 
-            public int getCurrent_css() {
+            public double getCurrent_css() {
                 return current_css;
             }
 
-            public void setCurrent_css(int current_css) {
+            public void setCurrent_css(double current_css) {
                 this.current_css = current_css;
             }
